+++ LPGManager/Interfaces/CompanyInterface/ICompanyService.cs
@@ -4,9 +4,9 @@ namespace LPGManager.Interfaces.CompanyInterface
     public interface ICompanyService
     {       
         Task<IEnumerable<Company>> GetAllAsync();
-        Task<Company> GetAsync(int id);
+        Task<Company> GetAsync(long id);
         Task<Company> AddAsync(Company company);
         Task<Company> UpdateAsync(Company model);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
     }
 }
