+++ LPGManager/Interfaces/PurchasesInterface/IPurchaseMasterService.cs
@@ -6,9 +6,9 @@ namespace LPGManager.Interfaces.PurchasesInterface
     public interface IPurchaseMasterService
     {
         Task<IEnumerable<PurchaseMaster>> GetAllAsync();        
-        Task<PurchaseMaster> GetAsync(int id);
+        Task<PurchaseMaster> GetAsync(long id);
         PurchaseMaster AddAsync(PurchaseMasterDtos model);
         Task<PurchaseMaster> UpdateAsync(PurchaseMaster model);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
     }
 }
