+++ LPGManager/Interfaces/SettingsInterface/IProductService.cs
@@ -5,9 +5,9 @@ namespace LPGManager.Interfaces.SettingsInterface
     public interface IProductService
     {
         Task<IEnumerable<Product>> GetAllAsync();
-        Task<Product> GetAsync(int id);
+        Task<Product> GetAsync(long id);
         Task<Product> AddAsync(Product supplier);
         Task<Product> UpdateAsync(Product model);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
     }
 }
