+++ LPGManager/Interfaces/SettingsInterface/IProductTypeService.cs
@@ -5,9 +5,9 @@ namespace LPGManager.Interfaces.SettingsInterface
     public interface IProductTypeService
     {
         Task<IEnumerable<ProductType>> GetAllAsync();
-        Task<ProductType> GetAsync(int id);
+        Task<ProductType> GetAsync(long id);
         Task<ProductType> AddAsync(ProductType productType);
         Task<ProductType> UpdateAsync(ProductType model);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
     }
 }
