+++ LPGManager/Interfaces/SettingsInterface/ISizeService.cs
@@ -5,9 +5,9 @@ namespace LPGManager.Interfaces.SettingsInterface
     public interface ISizeService
     {
         Task<IEnumerable<Size>> GetAllAsync();
-        Task<Size> GetAsync(int id);
+        Task<Size> GetAsync(long id);
         Task<Size> AddAsync(Size size);
         Task<Size> UpdateAsync(Size model);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
     }
 }
